Move filter search view to planned maintenance
