+++ src/argus/htmx/incident/views.py
@@ -7,10 +7,9 @@ from urllib.parse import urlencode
 from typing import Optional, Any
 
 from django import forms
-from django.contrib.auth import get_user_model
 from django.contrib import messages
+from django.contrib.auth import get_user_model
 from django.core.paginator import Paginator
-from django.db.models import Q
 from django.http import (
     HttpResponse,
     HttpResponseBadRequest,
@@ -18,10 +17,11 @@ from django.http import (
     JsonResponse,
     QueryDict,
 )
-from django_htmx.http import HttpResponseClientRefresh, retarget
-from django.utils.timezone import now as tznow
 from django.shortcuts import render, get_object_or_404
+from django.utils.timezone import now as tznow
 from django.views.decorators.http import require_POST, require_GET
+
+from django_htmx.http import HttpResponseClientRefresh, retarget
 
 from argus.incident.models import Incident, Tag
 from argus.incident.ticket.utils import get_ticket_plugin_path
@@ -265,20 +265,6 @@ def search_tags(request):
         tags = Tag.objects.filter(key__icontains=query)[:20]
 
     options = [{"id": str(tag), "text": str(tag)} for tag in tags]
-
-    return JsonResponse({"results": options})
-
-
-@require_GET
-def search_filters(request):
-    query = request.GET.get("q", "")
-
-    filters = Filter.objects.select_related("user")
-    if query:
-        filters = filters.filter(Q(name__icontains=query) | Q(user__username__icontains=query))
-    filters = sorted(filters, key=lambda f: (f.user != request.user, f.name))[:20]
-
-    options = [{"id": f.pk, "text": f"{f.name} ({f.user.username})"} for f in filters]
 
     return JsonResponse({"results": options})
